backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
joblib
scikit-learn==1.6.1
xgboost
catboost
lightgbm
shap
python-multipart
openpyxl
sqlalchemy
psycopg2-binary
passlib[argon2]
python-jose[cryptography]
celery
redis
google-generativeai
authlib
httpx
google-auth
